programs/ocrscripts/ocrscripts/requirements_ocr_local.txt

pytesseract
tesserocr
easyocr
paddleocr>=2.6
python-doctr[torch]
keras-ocr
mmocr
